Hopefully you can create your diagram it help figure out the zoom

+++ app/src/main/java/com/example/pixelpicker1/Crosshair.java
@@ -2,8 +2,8 @@ package com.example.pixelpicker1;
 
 public class Crosshair {
 
-	int x; // the X coordinate
-	int y; // the Y coordinate
+	float x; // the X coordinate
+	float y; // the Y coordinate
 
 	public Crosshair(int x, int y) {
 		this.x = x;
